+++ packages/core/src/metrics/aggregator.ts
@@ -1,0 +1,9 @@
+
+class MetricsAggregator {
+  private _metrics: Map<string, Metric>;
+
+  public constructor() {
+    this._metrics = new Map();
+  }
+}
+
